+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # CHANGELOG
 
+# Version 0.0.7
+
+- fixed issue: Parameters as symbols
+- added language contribution
+
 # Version 0.0.6
 
 - fixed issue: Multi-Dim support